Add logo to nav bar
